Extract TaskHeader subcomponent
